cambios en productos completado

+++ app_con_datos/lib/settings/database_connection.dart
@@ -43,7 +43,7 @@ class DatabaseConnection {
           descripcion TEXT,
           precio REAL NOT NULL,
           costo REAL NOT NULL,
-          stock INTEGER NOT NULL,
+          stock REAL NOT NULL,
           categoriaId INTEGER,
           FOREIGN KEY (categoriaId) REFERENCES categories(id)
         );
